Extract 'ferrix' into another module; add copyrights
